requirements.txt
Flask
gunicorn
google-genai
python-dotenv
pandas
openpyxl